restructure seeding utils in to a separate tools package
