+++ cmd/server/main.go
@@ -8,6 +8,7 @@ import (
 
 	"github.com/broadinstitute/sherlock/internal/db"
 	"github.com/broadinstitute/sherlock/internal/sherlock"
+	"github.com/broadinstitute/sherlock/internal/tools"
 	_ "github.com/golang-migrate/migrate/v4/database/postgres"
 	_ "github.com/golang-migrate/migrate/v4/source/file"
 	_ "github.com/jackc/pgx/v4/stdlib"
@@ -29,9 +30,10 @@ func main() {
 	if err := db.ApplyMigrations("db/migrations"); err != nil {
 		log.Println(err)
 	}
-	// if _, err := db.SeedServices(dbConn); err != nil {
-	// 	log.Println(err)
-	// }
+
+	if _, err := tools.SeedServices(dbConn); err != nil {
+		log.Println(err)
+	}
 
 	log.Println("starting sherlock server")
 	app := sherlock.New(dbConn)
