+++ internal/tools/seed.go
@@ -1,4 +1,4 @@
-package db
+package tools
 
 import (
 	"fmt"
